refactor(services): extract HTTP client utilities from subgraph service
Task: Extract HTTP-specific utility functions from the monolithic
subgraph_service.rs into dedicated modules to improve code organization
and maintainability.

Why: The subgraph service contains ~1600 lines mixing multiple concerns
including HTTP utilities, batching, WebSocket handling, and core service
logic. This extraction improves separation of concerns and testability
while laying groundwork for future HTTP protocol enhancements.

How: Created apollo-router/src/services/subgraph_impl/http/ module structure
and moved HTTP client utilities:
- get_uri_details() for URI parsing (host, port, path extraction)
- http_response_to_graphql_response() for response conversion
- get_graphql_content_type() for content type validation
- do_fetch() for core HTTP client execution
- ContentType enum and HTTP constants

Technical details: Maintained backward compatibility by re-exporting
APPLICATION_JSON_HEADER_VALUE. No behavioral changes - functions moved
without modification. Compatible with existing content negotiation plugin
(different scopes: router vs subgraph level).

Review strategy: Verify that all HTTP utilities in subgraph_impl/http/client.rs
maintain identical function signatures and behavior. Check that imports in
subgraph_service.rs correctly reference the new module paths. Confirm that
APPLICATION_JSON_HEADER_VALUE re-export maintains compatibility for existing
consumers in http_ext.rs and authentication/mod.rs.
