+++ apollo-router/src/services/mod.rs
@@ -6,6 +6,7 @@ use parking_lot::Mutex;
 
 pub(crate) use self::execution::service::*;
 pub(crate) use self::query_planner::*;
+pub(crate) use self::subgraph_service::APPLICATION_JSON_HEADER_VALUE;
 pub(crate) use self::subgraph_service::*;
 pub(crate) use self::supergraph::service::*;
 use crate::graphql::Request;
@@ -43,6 +44,7 @@ pub(crate) mod new_service;
 pub(crate) mod query_planner;
 pub mod router;
 pub mod subgraph;
+pub(crate) mod subgraph_impl;
 pub(crate) mod subgraph_service;
 pub mod supergraph;
