+++ apollo-router/src/services/subgraph_impl/http/mod.rs
@@ -1,0 +1,10 @@
+//! HTTP client functionality for subgraph communication
+
+pub(crate) mod client;
+
+pub(crate) use client::ACCEPT_GRAPHQL_JSON;
+pub(crate) use client::APPLICATION_JSON_HEADER_VALUE;
+pub(crate) use client::ContentType;
+pub(crate) use client::do_fetch;
+pub(crate) use client::get_uri_details;
+pub(crate) use client::http_response_to_graphql_response;
