+++ apollo-router/src/services/subgraph_impl/mod.rs
@@ -1,0 +1,3 @@
+//! Subgraph service modules
+
+pub(crate) mod http;
